+++ example/android/app/src/main/java/com/groovinchip/flutter/groovinwidgetsexample/MainActivity.java
@@ -1,13 +1,7 @@
 package com.groovinchip.flutter.groovinwidgetsexample;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
+
 }