+++ lib/src/widgets/groovin_expansion_tile.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide ListTile, ListTileTheme;
 import 'package:flutter/widgets.dart';
 
 import 'groovin_list_tile.dart';
@@ -7,49 +7,65 @@ const Duration _kExpand = Duration(milliseconds: 200);
 
 /// A slightly more customizable [ExpansionTile] widget.
 ///
-/// It allows for a subtitle, customizing the [BoxDecoration], and the border
+/// It allows for customizing the [BoxDecoration] and the border
 /// radius of the [InkWell] of the [ListTile] that the [ExpansionTile] creates.
 class GroovinExpansionTile extends StatefulWidget {
   /// Creates a [ListTile] with a trailing button that expands or collapses
   /// the tile to reveal or hide the [children]. The [initiallyExpanded] property must
   /// be non-null.
   const GroovinExpansionTile({
-    Key key,
+    Key? key,
     this.leading,
-    @required this.title,
+    required this.title,
     this.subtitle,
     this.backgroundColor,
     this.defaultTrailingIconColor = Colors.grey,
-    this.onExpansionChanged,
-    this.boxDecoration,
-    this.inkwellRadius,
     this.children = const <Widget>[],
     this.trailing,
     this.initiallyExpanded = false,
-  })  : assert(initiallyExpanded != null),
+    this.maintainState = false,
+    this.tilePadding,
+    this.expandedCrossAxisAlignment,
+    this.expandedAlignment,
+    this.childrenPadding,
+    this.collapsedBackgroundColor,
+    this.onExpansionChanged,
+    this.boxDecoration,
+    this.inkwellRadius,
+  })  : assert(
+          expandedCrossAxisAlignment != CrossAxisAlignment.baseline,
+          'CrossAxisAlignment.baseline is not supported since the expanded children '
+          'are aligned in a column, not a row. Try to use another constant.',
+        ),
         super(key: key);
 
   /// A widget to display before the title.
-  final Widget leading;
+  ///
+  /// Typically a [CircleAvatar] widget.
+  final Widget? leading;
 
   /// The primary content of the list item.
+  ///
+  /// Typically a [Text] widget.
   final Widget title;
 
-  /// Optional content to display below the title
-  final Widget subtitle;
+  /// Additional content displayed below the title.
+  ///
+  /// Typically a [Text] widget.
+  final Widget? subtitle;
 
   /// BoxDecoration for this widget
-  final BoxDecoration boxDecoration;
+  final BoxDecoration? boxDecoration;
 
   /// Represents the radius of the corners of the InkWell used by GroovinListTile
-  final BorderRadius inkwellRadius;
+  final BorderRadius? inkwellRadius;
 
   /// Called when the tile expands or collapses.
   ///
   /// When the tile starts expanding, this function is called with the value
   /// true. When the tile starts collapsing, this function is called with
   /// the value false.
-  final ValueChanged<bool> onExpansionChanged;
+  final ValueChanged<bool>? onExpansionChanged;
 
   /// The widgets that are displayed when the tile expands.
   ///
@@ -57,16 +73,72 @@ class GroovinExpansionTile extends StatefulWidget {
   final List<Widget> children;
 
   /// The color to display behind the sublist when expanded.
-  final Color backgroundColor;
+  final Color? backgroundColor;
+
+  /// When not null, defines the background color of tile when the sublist is collapsed.
+  final Color? collapsedBackgroundColor;
 
   /// The color to assign to the default trailing icon
-  final Color defaultTrailingIconColor;
+  final Color? defaultTrailingIconColor;
 
   /// A widget to display instead of a rotating arrow icon.
-  final Widget trailing;
+  final Widget? trailing;
 
   /// Specifies if the list tile is initially expanded (true) or collapsed (false, the default).
   final bool initiallyExpanded;
+
+  /// Specifies whether the state of the children is maintained when the tile expands and collapses.
+  ///
+  /// When true, the children are kept in the tree while the tile is collapsed.
+  /// When false (default), the children are removed from the tree when the tile is
+  /// collapsed and recreated upon expansion.
+  final bool maintainState;
+
+  /// Specifies padding for the [ListTile].
+  ///
+  /// Analogous to [ListTile.contentPadding], this property defines the insets for
+  /// the [leading], [title], [subtitle] and [trailing] widgets. It does not inset
+  /// the expanded [children] widgets.
+  ///
+  /// When the value is null, the tile's padding is `EdgeInsets.symmetric(horizontal: 16.0)`.
+  final EdgeInsetsGeometry? tilePadding;
+
+  /// Specifies the alignment of [children], which are arranged in a column when
+  /// the tile is expanded.
+  ///
+  /// The internals of the expanded tile make use of a [Column] widget for
+  /// [children], and [Align] widget to align the column. The `expandedAlignment`
+  /// parameter is passed directly into the [Align].
+  ///
+  /// Modifying this property controls the alignment of the column within the
+  /// expanded tile, not the alignment of [children] widgets within the column.
+  /// To align each child within [children], see [expandedCrossAxisAlignment].
+  ///
+  /// The width of the column is the width of the widest child widget in [children].
+  ///
+  /// When the value is null, the value of `expandedAlignment` is [Alignment.center].
+  final Alignment? expandedAlignment;
+
+  /// Specifies the alignment of each child within [children] when the tile is expanded.
+  ///
+  /// The internals of the expanded tile make use of a [Column] widget for
+  /// [children], and the `crossAxisAlignment` parameter is passed directly into the [Column].
+  ///
+  /// Modifying this property controls the cross axis alignment of each child
+  /// within its [Column]. Note that the width of the [Column] that houses
+  /// [children] will be the same as the widest child widget in [children]. It is
+  /// not necessarily the width of [Column] is equal to the width of expanded tile.
+  ///
+  /// To align the [Column] along the expanded tile, use the [expandedAlignment] property
+  /// instead.
+  ///
+  /// When the value is null, the value of `expandedCrossAxisAlignment` is [CrossAxisAlignment.center].
+  final CrossAxisAlignment? expandedCrossAxisAlignment;
+
+  /// Specifies padding for [children].
+  ///
+  /// When the value is null, the value of `childrenPadding` is [EdgeInsets.zero].
+  final EdgeInsetsGeometry? childrenPadding;
 
   @override
   _GroovinExpansionTileState createState() => _GroovinExpansionTileState();
@@ -84,11 +156,10 @@ class _GroovinExpansionTileState extends State<GroovinExpansionTile>
   final ColorTween _iconColorTween = ColorTween();
   final ColorTween _backgroundColorTween = ColorTween();
 
-  AnimationController _controller;
-  Animation<double> _iconTurns;
-  Animation<double> _heightFactor;
-  Animation<Color> _headerColor;
-  Animation<Color> _iconColor;
+  late AnimationController _controller;
+  late Animation<double> _iconTurns;
+  late Animation<double> _heightFactor;
+  late Animation<Color?> _iconColor;
 
   bool _isExpanded = false;
 
@@ -98,11 +169,10 @@ class _GroovinExpansionTileState extends State<GroovinExpansionTile>
     _controller = AnimationController(duration: _kExpand, vsync: this);
     _heightFactor = _controller.drive(_easeInTween);
     _iconTurns = _controller.drive(_halfTween.chain(_easeInTween));
-    _headerColor = _controller.drive(_headerColorTween.chain(_easeInTween));
     _iconColor = _controller.drive(_iconColorTween.chain(_easeInTween));
 
-    _isExpanded =
-        PageStorage.of(context)?.readState(context) ?? widget.initiallyExpanded;
+    _isExpanded = PageStorage.of(context)?.readState(context) as bool? ??
+        widget.initiallyExpanded;
     if (_isExpanded) _controller.value = 1.0;
   }
 
@@ -128,30 +198,24 @@ class _GroovinExpansionTileState extends State<GroovinExpansionTile>
       PageStorage.of(context)?.writeState(context, _isExpanded);
     });
     if (widget.onExpansionChanged != null)
-      widget.onExpansionChanged(_isExpanded);
+      widget.onExpansionChanged!(_isExpanded);
   }
 
-  Widget _buildChildren(BuildContext context, Widget child) {
-    final Color titleColor = _headerColor.value;
-
+  Widget _buildChildren(BuildContext context, Widget? child) {
     return Container(
       decoration: widget.boxDecoration,
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          IconTheme.merge(
-            data: IconThemeData(color: _iconColor.value),
+          ListTileTheme.merge(
+            iconColor: _iconColor.value,
+            //textColor: _headerColor.value,
             child: GroovinListTile(
-              onTap: _handleTap,
               inkwellRadius: widget.inkwellRadius,
+              onTap: _handleTap,
+              contentPadding: widget.tilePadding,
               leading: widget.leading,
-              title: DefaultTextStyle(
-                style: Theme.of(context)
-                    .textTheme
-                    .subtitle1
-                    .copyWith(color: titleColor),
-                child: widget.title,
-              ),
+              title: widget.title,
               subtitle: widget.subtitle,
               trailing: widget.trailing ??
                   RotationTransition(
@@ -165,6 +229,7 @@ class _GroovinExpansionTileState extends State<GroovinExpansionTile>
           ),
           ClipRect(
             child: Align(
+              alignment: widget.expandedAlignment ?? Alignment.center,
               heightFactor: _heightFactor.value,
               child: child,
             ),
@@ -177,24 +242,42 @@ class _GroovinExpansionTileState extends State<GroovinExpansionTile>
   @override
   void didChangeDependencies() {
     final ThemeData theme = Theme.of(context);
-    _borderColorTween..end = theme.dividerColor;
+    _borderColorTween.end = theme.dividerColor;
     _headerColorTween
-      ..begin = theme.textTheme.subtitle1.color
+      ..begin = theme.textTheme.subtitle1!.color
       ..end = theme.accentColor;
     _iconColorTween
       ..begin = theme.unselectedWidgetColor
       ..end = theme.accentColor;
-    _backgroundColorTween..end = widget.backgroundColor;
+    _backgroundColorTween
+      ..begin = widget.collapsedBackgroundColor
+      ..end = widget.backgroundColor;
     super.didChangeDependencies();
   }
 
   @override
   Widget build(BuildContext context) {
     final bool closed = !_isExpanded && _controller.isDismissed;
+    final bool shouldRemoveChildren = closed && !widget.maintainState;
+
+    final Widget result = Offstage(
+        child: TickerMode(
+          child: Padding(
+            padding: widget.childrenPadding ?? EdgeInsets.zero,
+            child: Column(
+              crossAxisAlignment: widget.expandedCrossAxisAlignment ??
+                  CrossAxisAlignment.center,
+              children: widget.children,
+            ),
+          ),
+          enabled: !closed,
+        ),
+        offstage: closed);
+
     return AnimatedBuilder(
       animation: _controller.view,
       builder: _buildChildren,
-      child: closed ? null : Column(children: widget.children),
+      child: shouldRemoveChildren ? null : result,
     );
   }
 }